refactor: MineBoardRequest dto 패키지로 이동
