+++ src/main/kotlin/minesweeper/view/ViewResolver.kt
@@ -1,7 +1,7 @@
 package minesweeper.view
 
 import minesweeper.domain.board.MineBoard
-import minesweeper.domain.board.MineBoardRequest
+import minesweeper.domain.board.dto.MineBoardRequest
 
 class ViewResolver {
     val inputMineBoardRequest: () -> MineBoardRequest = InputView::inputMineBoardRequest
